Ajout de l'exportation au fonctionnement general de l'application
