+++ CodeGen/Chat.cs
@@ -1,7 +1,7 @@
 
 public class Chat
 {
-	private string nom;
+	public string nom;
 	public string prenom;
 
 	public Chat(string nom, string prenom)
@@ -14,12 +14,12 @@ public class Chat
 		
 	}
 
-	private string FaireQQch(string nom1, string nom2)
+	public string FaireQQch(string nom1, string nom2)
 	{
 		
 	}
 
-	private void FaireQQchAutre(string nom1, string nom2)
+	public string FaireQQchAutre(string nom1, string nom2)
 	{
 		
 	}
