+++ CodeGen/IVendable.cs
@@ -1,0 +1,10 @@
+
+public interface IVendable
+{
+	public string nom { get; set; }
+	public string prenom { get; set; }
+
+	public string FaireQQch(string nom1, string nom2);
+	public string FaireQQchAutre(string nom1, string nom2);
+
+}
