+++ CodeGen/src/main/kotlin/com/codegen/codegen/PrincipaleControleur.kt
@@ -4,7 +4,9 @@ package com.codegen.codegen
 
 import com.codegen.codegen.composants.*
 import com.codegen.codegen.fichiersCode.Classe
+import com.codegen.codegen.fichiersCode.Interface
 import com.codegen.codegen.fichiersCode.MotCleeClasse
+import com.codegen.codegen.projet.Projet
 import javafx.application.Platform
 import javafx.fxml.FXML
 import javafx.fxml.FXMLLoader
@@ -14,6 +16,7 @@ import javafx.scene.control.*
 import javafx.scene.layout.VBox
 import javafx.stage.FileChooser
 import javafx.stage.Stage
+import java.io.File
 
 /**
  *  Classe du contrôleur de l'interface principale.
@@ -63,6 +66,15 @@ class PrincipaleControleur {
     @FXML
     private var zoneAffichageClasse: VBox = VBox()
 
+    /**
+     * Variable qui définissent le projet et la classe active
+     *
+     * @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+     */
+    private lateinit var projetActif: Projet
+
+    private lateinit var classeActive: Classe
+
 
     /**
      * Fonction d'initialisation des composants
@@ -89,12 +101,12 @@ class PrincipaleControleur {
         menuItemAjouterHeritage.setOnAction {
             surAjouterHeritage()
         }
-//        menuItemExporterClasse.setOnAction {
-//            surExporterClasse()
-//        }
-//        menuItemExporterClasse.setOnAction {
-//            surExporterProjet()
-//        }
+        menuItemExporterClasse.setOnAction {
+            surExporterClasse()
+        }
+        menuItemExporterProjet.setOnAction {
+            surExporterProjet()
+        }
         menuItemSupprimerClasse.setOnAction {
             surSupprimerClasse()
         }
@@ -106,19 +118,24 @@ class PrincipaleControleur {
 //        }
         // Définir ce qu'il y a dans la classe
         // @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
-        var propriete1 = Propriete(Visibilite.private, "string", "nom")
+        var propriete1 = Propriete(Visibilite.public, "string", "nom")
         var propriete2 = Propriete(Visibilite.public, "string", "prenom")
         var listeProprieteConstructeur = listOf(propriete1, propriete2)
+        var listeProprieteConstructeur2 = listOf(propriete1)
         var listePropriete = listOf(propriete1, propriete2)
-        var param1 = Parametre("string", "nom1");
-        var param2 = Parametre("string", "nom2");
-        var liste = listOf(param1, param2);
-        var uneMethode = Methode(Visibilite.private, "string", "FaireQQch", liste, true)
-        var methode2 = Methode(Visibilite.private, "string", "FaireQQch", liste, false)
+        var liste = listOf(Parametre("string", "nom1"), Parametre("string", "nom2"))
+        var uneMethode = Methode(Visibilite.public, "string", "FaireQQch", liste, true)
+        var methode2 = Methode(Visibilite.public, "string", "FaireQQchAutre", liste, false)
         var listeMethode = listOf(uneMethode, methode2)
         var constructeur = Constructeur(Visibilite.public, listeProprieteConstructeur)
-        var listeConstructeur = listOf(constructeur)
+        var constructeur2 = Constructeur(Visibilite.public, listeProprieteConstructeur2)
+        var listeConstructeur = listOf(constructeur, constructeur2)
         val classe = Classe(Visibilite.public, MotCleeClasse.classique, listePropriete, listeMethode, listeConstructeur)
+        val listeClasse = listOf(classe)
+        val uneInterface = Interface(Visibilite.public, listeProprieteConstructeur, listeMethode)
+        val listeInterfaces = listOf(uneInterface)
+        projetActif = Projet(listeClasse, listeInterfaces)
+        classeActive = projetActif.classes[0]
         zoneAffichageClasse.children.add(ControleurClasse(classe))
     }
 
@@ -246,8 +263,14 @@ class PrincipaleControleur {
      *
      * @author Alexandre del Fabbro - 2166311@etudiant.cegepvicto.ca
      */
-//    @FXML
-//    private fun surExporterClasse() {
+    @FXML
+    private fun surExporterClasse() {
+        // @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+        print("exporter classe")
+        exporterClasseCSharp(classeActive)
+        // Pour l'instant, la selection de dossier n'est pas encore intégré
+        // dans l'application. L'exportation se fait dans le dossier de projet
+        // Fin des ajouts de Cedric
 //        val selecteurFichier = FileChooser()
 //        selecteurFichier.title = "Choisissez un emplacement pour l'export de la classe"
 //        selecteurFichier.extensionFilters.add(
@@ -258,7 +281,21 @@ class PrincipaleControleur {
 //            val cheminExportClasse = fichierSelectionne.absolutePath
 //            // Exportez la classe vers le chemin cheminExportClasse
 //        }
-//    }
+    }
+
+    /**
+     * Fonction pour exporter le projet actif
+     *
+     * @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+     */
+    @FXML
+    private fun surExporterProjet() {
+        print("exporter projet")
+        exporterCSharp(projetActif)
+
+        // Pour l'instant, la selection de dossier n'est pas encore intégré
+        // dans l'application. L'exportation se fait dans le dossier de projet
+    }
 
     /**
      * Fonction pour la suppression d'une classe
@@ -310,5 +347,98 @@ class PrincipaleControleur {
             val nouveauNomInterface = resultat.get()
             // Ajoutez le nouvel interface avec le nom nouveauNomInterface
         }
+    }
+
+    /**
+     * Fonction qui exporte toutes les classes et interfaces
+     * du projet en C#
+     *
+     * @param projet le projet à exporter
+     *
+     * @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+     */
+    fun exporterCSharp(projet: Projet) {
+        for (classe in projet.classes) {
+            exporterClasseCSharp(classe)
+        }
+        for (uneInterface in projet.interfaces) {
+            exporterInterfaceCSharp(uneInterface)
+        }
+    }
+
+    /**
+     * Fonction qui exporte la classe en paramètres en C#
+     *
+     * @param classe la classe à exporter
+     *
+     * @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+     */
+    fun exporterClasseCSharp(classe: Classe) {
+        val className = "Chat"
+        val fileName = "$className.cs"
+        var fileContent = "\n" + classe.visibilite + " class " + className + "\n{\n"
+        for (propriete in classe.proprietes) {
+            fileContent += "\t" + propriete.visibilite + " " + propriete.type + " " + propriete.nom + ";\n"
+        }
+        fileContent += "\n"
+        for (constructeur in classe.constructeurs) {
+            fileContent += "\t" + constructeur.visibilite + " " + className + "("
+            for (index in constructeur.proprietes.indices) {
+                fileContent += constructeur.proprietes[index].type + " " + constructeur.proprietes[index].nom
+                if (index + 1 != constructeur.proprietes.size) {
+                    fileContent += ", "
+                }
+            }
+            fileContent += ")\n\t{\n\t\t\n\t}\n\n"
+        }
+        for (methode in classe.methodes) {
+            fileContent += "\t" + methode.visibilite + " " + methode.typeRetour + " " + methode.nom + "("
+            fileContent += ajouterParametresMethodeCSharp(methode)
+            fileContent += ")\n\t{\n\t\t\n\t}\n\n"
+        }
+        fileContent += "\n}"
+        File(fileName).writeText(fileContent)
+    }
+
+    /**
+     * Fonction qui exporte l'interface en paramètre en C#
+     *
+     * @param uneInterface l'interface à exporter
+     *
+     * @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+     */
+    fun exporterInterfaceCSharp(uneInterface: Interface) {
+        val nomInterface = "IVendable"
+        val fileName = "$nomInterface.cs"
+        var fileContent = "\n" + uneInterface.visibilite +  " interface $nomInterface\n{\n"
+        for (propriete in uneInterface.proprietes) {
+            fileContent += "\t" + propriete.visibilite + " " + propriete.type + " " + propriete.nom + " { get; set; }\n"
+        }
+        fileContent += "\n"
+        for (methode in uneInterface.methodes) {
+            fileContent += "\t" + methode.visibilite + " " + methode.typeRetour + " " + methode.nom + "("
+            fileContent += ajouterParametresMethodeCSharp(methode) + ");\n"
+        }
+        fileContent += "\n}"
+        File(fileName).writeText(fileContent)
+    }
+
+    /**
+     * Fonction qui ajoute les paramètres d'une méthode dans un string
+     * pour l'affichage lors de l'exportation
+     *
+     * @param methode la méthode dont on veut avoir un string de ses paramètres
+     *
+     * @author Cedric Garand - 2135500@etudiant.cegepvicto.ca
+     */
+    private fun ajouterParametresMethodeCSharp(methode: Methode): String {
+        var fileContent = ""
+        for (index in methode.parametres.indices) {
+            fileContent += methode.parametres[index].type + " " + methode.parametres[index].nom
+            if (index + 1 != methode.parametres.size) {
+                fileContent += ", "
+            }
+        }
+        return fileContent
     }
 }
